Create n_queen.c
This is the N-queens problem. 

In this problem, we are given with a matrix where the size is given as input. A queen in a chess board can't be placed where the queen has opponent in the 90 degree position or diagonal position. So we are finding the positions were the queen can be kept safely in the given matrix.